Collision Detection: Step Height Update
The logic for step height has been enhanced. However, I'm aware of a bug where the character, when positioned on the edge of the collider they are stepping on, may occasionally fall through it.

For now, I've made some adjustments, and it appears that the issue is occurring less frequently. I'll continue to address this in more detail at a later date.